+++ modules/convolution/tests/ConvolutionEngineTests.cpp
@@ -24,7 +24,7 @@ void  operator delete[] (void* p, std::size_t) noexcept { std::free (p); }
 
 using namespace felitronics;
 
-struct Lcg { unsigned long s; float next() { s = s * 6364136223846793005ULL + 1442695040888963407ULL; return (float) ((s >> 40) & 0xffff) / 32768.0f - 1.0f; } };
+struct Lcg { unsigned long long s; float next() { s = s * 6364136223846793005ULL + 1442695040888963407ULL; return (float) ((s >> 40) & 0xffff) / 32768.0f - 1.0f; } };
 
 static double maxDeriv (const std::vector<float>& y, int from, int to)
 {
