+++ modules/deesser/tests/DeEsserTests.cpp
@@ -83,7 +83,7 @@ int main()
     {
         DP p; p.mode = Mode::SplitBand; p.fc = 6000.0; p.scQ = 1.0; p.thresholdDb = 12.0; p.ratio = 4.0;   // never triggers
         DE d; d.prepare (sr, 512, 1); d.setParams (p);
-        unsigned long s = 4; auto rng = [&]() { s = s * 6364136223846793005ULL + 1442695040888963407ULL; return (float) ((s >> 40) & 0xffff) / 32768.0f - 1.0f; };
+        unsigned long long s = 4; auto rng = [&]() { s = s * 6364136223846793005ULL + 1442695040888963407ULL; return (float) ((s >> 40) & 0xffff) / 32768.0f - 1.0f; };
         const int N = 8000; std::vector<float> x (N), y (N);
         for (int i = 0; i < N; ++i) { x[i] = 0.2f * rng(); y[i] = x[i]; }
         for (int o = 0; o < N; o += 512) { float* io[1] { y.data() + o }; d.process (io, 1, std::min (512, N - o)); }
@@ -143,7 +143,7 @@ int main()
         DP p; p.mode = Mode::SplitBand; p.fc = 6000.0; p.scQ = 1.0; p.thresholdDb = 12.0; p.ratio = 4.0;   // never triggers
         DE d; d.prepare (sr, 512, 1); d.setParams (p);
         eq::Crossover2 ref; ref.prepare (sr, 1); ref.setFrequency (6000.0f);
-        unsigned long s = 6; auto rng = [&]() { s = s * 6364136223846793005ULL + 1442695040888963407ULL; return (float) ((s >> 40) & 0xffff) / 32768.0f - 1.0f; };
+        unsigned long long s = 6; auto rng = [&]() { s = s * 6364136223846793005ULL + 1442695040888963407ULL; return (float) ((s >> 40) & 0xffff) / 32768.0f - 1.0f; };
         const int N = 6000; std::vector<float> x (N), y (N);
         for (int i = 0; i < N; ++i) { x[i] = 0.2f * rng(); y[i] = x[i]; }
         for (int o = 0; o < N; o += 512) { float* io[1] { y.data() + o }; d.process (io, 1, std::min (512, N - o)); }
